+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft Sendcloud Plugin
 
+## 2.2.0 - UNRELEASED
+
+### Added
+- Added an option to configure the Sendcloud order number ([#1](https://github.com/white-nl/commerce-sendcloud/issues/1))
+
 ## 2.1.3 - 2023-04-04
 
 ### Changed